+++ src/FastNoiseSIMD_sse41.cpp
@@ -33,7 +33,7 @@
 // Depending on the compiler this file may need to have SSE4.1 code generation compiler flags enabled
 #ifdef FN_COMPILE_SSE41
 #define SIMD_LEVEL_H FN_SSE41
-#include "FastNoiseSIMD/FastNoiseSIMD_internal.h"
+#include "FastNoiseSIMD_internal.h"
 #include <smmintrin.h> //SSE4.1
 
 #define SIMD_LEVEL FN_SSE41
